+++ src/components/CardPreview/CardPreview.tsx
@@ -3,10 +3,15 @@
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { formatLayoutLabel } from "@/lib/svg/languages/utils";
 import { themes } from "@/lib/themes/configs/registry";
-import { BADGE_STYLES } from "@/lib/svg/badge/configs/registry";
-import { renderBadge } from "@/lib/svg/badge";
 import { LANG_CHART_LAYOUTS } from "@/lib/types";
 import { CardOpts, LangOpts, MiniOpts, SparkOpts } from "./types";
+import {
+  appendCardParams,
+  appendCardOrderParams,
+  appendLangsParams,
+  appendMiniParams,
+  appendSparklineParams,
+} from "./utils/embedUrl";
 import {
   CARD_DEFAULTS,
   LANG_DEFAULTS,
@@ -15,122 +20,8 @@ import {
 } from "./configs";
 import { STAT_OPTIONS, EMBED_LABELS, MINI_METRICS } from "./configs";
 import { EmbedType } from "./types";
-
-function BadgeStylePicker({
-  value,
-  onChange,
-}: {
-  value: string;
-  onChange: (s: string) => void;
-}) {
-  return (
-    <div className="mt-2 grid grid-cols-2 sm:grid-cols-3 gap-2">
-      {BADGE_STYLES.map((opt) => {
-        const isActive = value === opt.key;
-        const svg = renderBadge("STYLE", "preview", "58a6ff", opt.key);
-        return (
-          <button
-            key={opt.key}
-            type="button"
-            onClick={() => onChange(opt.key)}
-            aria-pressed={isActive}
-            className={`group flex flex-col items-center gap-1.5 rounded-lg border p-2.5 transition-all duration-150 ${
-              isActive
-                ? "border-[#58a6ff] bg-[#58a6ff]/10 shadow-[0_0_0_1px_rgba(88,166,255,0.25)]"
-                : "border-[#30363d] bg-[#161b22] hover:border-[#484f58]"
-            }`}
-          >
-            <span
-              aria-hidden
-              className="flex h-7 items-center"
-              dangerouslySetInnerHTML={{ __html: svg }}
-            />
-            <span
-              className={`text-[10px] font-semibold uppercase tracking-wider transition-colors ${
-                isActive
-                  ? "text-[#58a6ff]"
-                  : "text-[#8b949e] group-hover:text-[#c9d1d9]"
-              }`}
-            >
-              {opt.label}
-            </span>
-          </button>
-        );
-      })}
-    </div>
-  );
-}
-
-function usePatchState<T>(initial: T) {
-  const [state, setState] = useState<T>(initial);
-  const patch = useCallback((p: Partial<T>) => {
-    setState((prev) => ({ ...prev, ...p }));
-  }, []);
-  return [state, patch] as const;
-}
-
-function appendCardParams(p: URLSearchParams, card: CardOpts) {
-  if (!card.showIcons) p.set("show_icons", "false");
-  if (!card.showRing) p.set("show_ring", "false");
-  if (card.hideBorder) p.set("hide_border", "true");
-  if (card.hideTitle) p.set("hide_title", "true");
-  if (card.borderRadius !== "4.5") p.set("border_radius", card.borderRadius);
-  if (card.customTitle.trim()) p.set("custom_title", card.customTitle.trim());
-  if (card.size === "compact") {
-    p.set("size", "compact");
-    if (card.compactCount !== 6)
-      p.set("compact_count", String(card.compactCount));
-    if (card.showEmoji) p.set("show_emoji", "true");
-  }
-}
-
-function appendCardOrderParams(
-  p: URLSearchParams,
-  statsOrder: string[],
-  hiddenStats: string[],
-) {
-  if (hiddenStats.length > 0) p.set("hide", hiddenStats.join(","));
-
-  const visibleOrder = statsOrder.filter((k) => !hiddenStats.includes(k));
-  const defaultVisible = STAT_OPTIONS.map((s) => s.key).filter(
-    (k) => !hiddenStats.includes(k),
-  );
-  const isReordered =
-    visibleOrder.length === defaultVisible.length &&
-    visibleOrder.some((k, i) => k !== defaultVisible[i]);
-  if (isReordered) p.set("order", visibleOrder.join(","));
-}
-
-function appendLangsParams(
-  p: URLSearchParams,
-  langs: LangOpts,
-  customTitle: string,
-) {
-  if (langs.hideBorder) p.set("hide_border", "true");
-  if (langs.hideTitle) p.set("hide_title", "true");
-  if (langs.borderRadius !== "4.5") p.set("border_radius", langs.borderRadius);
-  if (customTitle.trim()) p.set("custom_title", customTitle.trim());
-  if (langs.maxLangs !== 8) p.set("max_langs", String(langs.maxLangs));
-  if (langs.layout !== "bar") p.set("layout", langs.layout);
-}
-
-function appendMiniParams(p: URLSearchParams, mini: MiniOpts) {
-  if (mini.metric !== "stars") p.set("metric", mini.metric);
-  if (mini.label.trim()) p.set("label", mini.label.trim());
-  if (mini.color.trim()) p.set("color", mini.color.trim());
-  if (mini.style !== "flat") p.set("style", mini.style);
-}
-
-function appendSparklineParams(p: URLSearchParams, spark: SparkOpts) {
-  p.set("days", spark.days || "30");
-  p.set("width", spark.width || "320");
-  p.set("height", spark.height || "80");
-  if (spark.hideBorder) p.set("hide_border", "true");
-  if (spark.borderRadius !== "6") p.set("border_radius", spark.borderRadius);
-  if (spark.lineColor.trim()) p.set("line_color", spark.lineColor.trim());
-  if (spark.fillColor.trim()) p.set("fill_color", spark.fillColor.trim());
-  if (spark.title.trim()) p.set("title", spark.title.trim());
-}
+import { usePatchState } from "./utils/state";
+import BadgeStylePicker from "./BadgeStylePicker";
 
 export default function CardPreview() {
   const [embedType, setEmbedType] = useState<EmbedType>("card");
